FIX UZS overrides for a-candy-girls-christmas, a-fairy-tale, asuka-x-samurai, bangkok-fighter, battlemech-mars, cloud9, cookoff-champion, dragon-myst, drift, fruity-verse, museum, pearl-fisher, seasons, seasons-autumn, seasons-spring, seasons-summer, seasons-winter, sky-jewels, sparta, valley-of-kings

+++ internal/engine/helpers.go
@@ -174,7 +174,7 @@ func (num Fixed) Pow(exp int) Fixed {
 }
 
 func NewFixedFromFloat(num float32) Fixed {
-	return Fixed(num * float32(fixedExp))
+	return Fixed(float64(num) * float64(fixedExp))
 }
 
 func NewFixedFromInt(num int) Fixed {
